Javaの`initLogger`のnative実装を移動
