+++ crates/voicevox_core_java_api/src/lib.rs
@@ -1,5 +1,6 @@
 mod common;
 mod info;
+mod logger;
 mod open_jtalk;
 mod synthesizer;
 mod user_dict;
